+++ lcapy/fexpr.py
@@ -64,7 +64,7 @@ class FourierDomainExpression(FourierDomain, Expr):
         return self.inverse_fourier(**assumptions)
 
     def norm_fourier(self, **assumptions):
-        """Convert to normalised Fourier domain."""
+        """Convert to normalized Fourier domain."""
         from .symbols import F
         from .dsym import dt
         
@@ -79,7 +79,7 @@ class FourierDomainExpression(FourierDomain, Expr):
         return result
 
     def norm_angular_fourier(self, **assumptions):
-        """Convert to normalised angular Fourier domain."""
+        """Convert to normalized angular Fourier domain."""
         from .symbols import Omega
         from .dsym import dt
         
@@ -115,7 +115,7 @@ class FourierDomainExpression(FourierDomain, Expr):
         `ylabel2` - the second y-axis label if needed, say for mag and phase
         `xscale` - the x-axis scaling, say for plotting as ms
         `yscale` - the y-axis scaling, say for plotting mV
-        `norm` - use normalised frequency
+        `norm` - use normalized frequency
         in addition to those supported by the matplotlib plot command.
         
         The plot axes are returned.  This is a tuple for magnitude/phase or